Edited functions from unsigned int to size_t if required, pre-created empty part 2 libft functions

+++ done/ft_calloc.c
@@ -10,10 +10,9 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stdlib.h>
 #include "libft.h"
 
-void	*ft_calloc(unsigned int nelem, unsigned int elsize)
+void	*ft_calloc(size_t nelem, size_t elsize)
 {
 	void	*new_array;
